fix(formulas): set newly created code inside formulas as empty
Fixes an issue where the code for the newly created formulas was
getting populated from the code for already opened formulas.

The problem occurred because the 'code' variable was propagating 'undefined'
to the code editor whenever a new formula was created. To address this, we
have added a check that verifies if the 'code' variable is 'undefined',
and if so, sets the value in the code editor to an empty string, ensuring
accurate calculations and proper initialization of new formulas.

Jira Issue: DS-509